completion of backend for EarthNbN

+++ backend/routes/api/bookings.js
@@ -3,9 +3,10 @@ const { Booking, User, Spot } = require('../../db/models');
 const { requireAuth } = require("../../utils/auth.js");
 
 const router = express.Router();
+const today = new Date()
 
-//READ **INCOMPLETE** - all current user bookings
-//missing previewImg on Spot
+
+//READ - all current user bookings
 router.get('/current', requireAuth, async (req, res, next) => {
     try {
         let reqUser = parseInt(req.user.id)
@@ -16,8 +17,7 @@ router.get('/current', requireAuth, async (req, res, next) => {
     }
 })
 
-//UPDATE **INCOMPLETE** - booking based on booking Id
-//Major Problems
+//UPDATE - booking based on booking Id
 router.put('/:bookingId', requireAuth, async (req, res, next) => {
     try {
         let { startDate, endDate } = req.body
@@ -48,23 +48,28 @@ router.put('/:bookingId', requireAuth, async (req, res, next) => {
         res.json(updateBooking)
     } catch (err) {
         if(err.message === "Booking couldn't be found"){err.status = 404}
-        else{err.status = 404}
+        else{err.status = 403}
         next(err)
     }
 })
 
-//DELETE **COMPLETE** - booking by booking ID
+//DELETE - booking by booking ID
 router.delete('/:bookingId', requireAuth, async (req, res, next) => {
     try {
         const bookingId = parseInt(req.params.bookingId)
+        if(isNaN(bookingId)){throw new Error("Your bookingId is not a number")}
         let findBooking = await Booking.findByPk(bookingId)
         if(!findBooking){throw new Error("Booking couldn't be found")}
         if(findBooking.userId !== req.user.id){throw new Error('Forbidden')}
+        let bookStart = findBooking.dataValues.startDate
+        if(bookStart.toISOString().slice(0, 10) <= today.toISOString().slice(0, 10)){throw new Error("Bookings that have been started can't be deleted")}
 
         await Booking.destroy({where: {id: bookingId}})
         res.json({ message: "Successfully deleted" })
     } catch (err) {
         if(err.message === "Booking couldn't be found"){err.status = 404}
+        if(err.message === "Your bookingId is not a number"){err.status = 404}
+        if(err.message === "Bookings that have been started can't be deleted"){err.status = 403}
         next(err)
     }
 })
